app/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
google-cloud-logging==2.6.0
google-cloud-firestore==2.3.0
google-cloud-storage==1.42.0
google-cloud-translate==3.4.0
google-cloud-language==2.2.2
